The rate would not work correctly if the value was a floating point number, fixed it so that it would work.

+++ ch01/exercises/simple.py
@@ -10,8 +10,8 @@ print(10 % 10)
 print(0 % 10)
 ## print(10 / 15)
 ## The last answer is repeating which causes high cpu usage.
-rate = int(input("What is the current exchange rate for Euro to USD? "))
-amount = int(input("How much Euro would you like to be exchanged for USD? "))
+rate = float(input("What is the current exchange rate for Euro to USD? "))
+amount = float(input("How much Euro would you like to be exchanged for USD? "))
 total = amount * rate
 result = total - 3
 print("We will be taking $3 for a service fee. Here is " + str(result) + " USD")
